Update Invidious Structures.swift

+++ Sources/Invidious-Swift/API Structs/Invidious Structures.swift
@@ -348,7 +348,7 @@ public struct InvRecommendedVideo: Codable {
     public let author, authorURL, authorID: String
     public let lengthSeconds: Int
     public let viewCountText: String
-    public let viewCount: Int
+    public let viewCount: Double
 
     public enum CodingKeys: String, CodingKey {
         case videoID = "videoId"
